requirements.txt
pydantic>=2.0
rich>=13.0
httpx>=0.25
GitPython>=3.1
